style: projects layout fixed
