--- projects/ngx-xchange-ui/src/components/projects-card/projects-card.component.html
@@ -1,26 +1,0 @@
-<button *ngFor="let project of projects" xcCardStyles [padding]="16" [gap]="16" (click)="openModal(project)">	
-    <div class="image-wrapper">
-        <img src={{project.image}} alt="project-background" class="project-image">
-    </div>
-    <div class="project-title">
-        <h6>{{project.title}}</h6>
-        <span  class="project-tags">
-            <div xcProfileTags="gray" *ngFor="let tags of project.tags">{{tags}}</div>
-        </span>
-    </div>
-    <div class="project-progress" *ngIf="metrics && objectives">
-        <xc-project-progress-bar [completed]="project.completed" [open]="project.open" [suggested]="project.suggested" [objectivesInProject]="50" class="project-progress-bar"/>
-    </div>
-    <div class="project-description">
-        <p>
-            {{project.description}}
-        </p>
-    </div>
-</button>
-
-<dialog class="modal-overlay" *ngIf="selectedProject" >
-    <app-project-modal [project]="selectedProject" class="modal" (close)="onCloseModal()"/>
-    <button class="close-modal-button" (click)="onCloseModal()">
-        <xc-icon-close-s class="close-modal-icon"></xc-icon-close-s>
-      </button>
-</dialog>
