+++ src/app/pages/projects/projects-card/projects-card.component.css
@@ -1,3 +1,13 @@
+.wrapper{
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  padding: 16px;
+  width: 100%;
+  border-radius: 16px;
+  background: white;
+}
+
 .image-wrapper {
   width: 100%;
   height: 200px;
