+++ src/app/pages/projects/projects.component.scss
@@ -1,15 +1,15 @@
 .container {
   display: flex;
   flex-direction: column;
-  width: 100%;
+  width: 100vw;
   height: 100%;
   overflow: hidden;
   &_content {
     display: flex;
     flex-direction: column;
-    gap: 30px;
+    gap: 25px;
     width: 100%;
-    height: calc(100% - 20px);
+    height: 100%;
     &-nav {
       display: flex;
       align-items: center;
@@ -26,7 +26,6 @@
           background: var(--xc-gradient-light-grey-on-white);
           padding: 8px 16px;
           transition: background 0.3s, color 0.3s;
-
           &.unselected {
             color: gray;
             background: var(--xc-light-grey);
@@ -38,19 +37,21 @@
       display: flex;
       flex-direction: column;
       gap: 21px;
-      flex: 1;
+      width: 100%;
+      height: 92%;
       &-wrapper {
-        height: calc(100vh - 200px);
+        align-items: start;
+        box-sizing: border-box;
+        display: grid;
+        grid-template-columns: repeat(auto-fill, minmax(490px, 1fr));
+        grid-template-rows: repeat(auto-fill, minmax(410px, 1fr));
+        gap: 32px;
         overflow-y: auto;
-        display: flex;
-        flex-direction: column;
-        gap: 16px;
-        &-injector {
-          display: grid;
-          grid-template-columns: repeat(auto-fill, minmax(400px, 1fr));
-          gap: 24px;
-          width: 100%;
-          padding-bottom: 20px;
+        scrollbar-width: none;
+        margin-bottom: 36px;
+        -ms-overflow-style: none;      
+        &::-webkit-scrollbar { 
+          display: none;
         }
       }
     }
@@ -62,15 +63,25 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  max-width: 480px;
-  min-height: 410px;
+  width: 490px;
+  height: 410px;
+  border-radius: 8px;
   background: white;
   padding: 16px;
-  gap: 16px;
-  &-icon {
-    width: 80px;
-    height: 80px;
-    stroke: var(--xc-black500);
+  &-content{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    background: var(--xc-grey200);
+    &-icon {
+      width: 80px;
+      height: 80px;
+      stroke: var(--xc-black500);
+      stroke-width: 1px;
+    }
   }
 }
 
@@ -80,7 +91,7 @@
   top: 0;
   left: 0;
   width: 100%;
-  height: 100%;
+  min-height: 100%;
   justify-content: center;
   align-items: center;
   z-index: 10;
