+++ src/app/pages/projects/stepper/stepper.component.html
@@ -101,7 +101,7 @@
       <div class="second-container-templates">
         <xc-projects-card
           class="second-container-templates-injector"
-          [metrics]="false"
+          
         />
       </div>
     </article>
